feat!: hardwareConcurrency sang u16 va cat theo he dieu hanh — 26,4% khoi luong dang mat vao unwrap_or(4)
THAY DOI PHA VO thu hai cua 0.3.0 (chua phat hanh, nen gop cung ban de nguoi
dung ngoai chiu dung mot lan vo). `hardware_concurrency`: u8 -> u16.

Cung lop loi voi VEIL-699 nhung nang hon 150 lan, va te hon ve chat:

  maxTouchPoints        khoi luong > u8 =  0.00%   mat thanh None
  hardwareConcurrency   khoi luong > u8 = 26.36%   mat thanh .unwrap_or(4)

`None` con noi duoc "khong co du lieu". `4` thi khong — no la mot con so TRONG
HOP LY, khong ai phan biet duoc voi mot may 4 nhan that.

NOI KIEU MOT MINH LA SAI, va phep do noi ro vi sao. Voi moi gia tri vuot u8, he
dieu hanh cua UA di kem:

  384   Linux=72.00   macOS=46.24   Windows=8.00
  448/512/640         ~0.04 tong cong — gan nhu khong ton tai

384 luong tren server Linux la CO THAT (EPYC 9754 hai socket = 256 nhan / 512
luong). 384 nhan tren mot may Mac thi KHONG ton tai. Nen phat nguyen gia tri se
tao mot dau van tay MOI, te hon dau van tay sai no dang thay.

Nguoi dung chon duong (2): noi kieu VA cat theo he dieu hanh.

  macOS     24   Mac Pro M2 Ultra
  iOS       10   iPad Pro M4
  Android   16   TRAN CO BIEN, khong phai so nhan cua mot may cu the
  Windows        khong cat — server that dat toi hang tram nhan
  Linux          khong cat — cung ly do

DOI CA HANH VI CO SAN, khong chi khoi phuc du lieu mat: truoc ban nay ho so iOS
va Android da nhan toi 186-192 nhan, vi u8 cho lot moi gia tri 25-255. Noi kieu
chi PHOI RA loi kha di do chu khong tao ra no. Do tren 3000 ho so:

             truoc                sau
  Ios        max 640, 20.4% >24   max 10
  Android    max 512, 19.0% >24   max 16
  MacOs      max 384              max 24
  Linux      max 144              max 144   khong doi
  Windows    max 640              max 640   khong doi

GOC CAT CO Y, khai kem tran: moi ho so macOS vuot 24 deu ve DUNG 24, nen gia tri
24 bi don cao hon tu nhien. Van la mot dau vet, chi nho hon dau vet cu (moi thu
ve 4). Nang cap khi co ly do that: rut lai tu phan phoi co dieu kien cua chinh
nut do thay vi cat cung.

BA BAI MOI, hai chay o trang thai DO truoc khi sua:

  mang khai 62 gia tri cho hardwareConcurrency nhung 4 gia tri khong doc lai
  duoc: ["*STRINGIFIED*384", "*STRINGIFIED*640", "*STRINGIFIED*512",
  "*STRINGIFIED*448"]

  ho so MacOs nhan 48 nhan sau khi loc con 48 — vuot tran 24

Bai thu ba la bai DAU-CUOI, va no ton tai vi mot ly do do duoc: ban sua dau tien
cua toi CHI loc macOS, moi bai don vi deu xanh, va chinh phep do dau-cuoi moi lo
ra iOS con nhan 640 nhan va Android 512. Dot bien xac nhan khoang mu do:

  go loi goi clamp khoi duong sinh, giu nguyen ham
    he_di_dong_...          ok       <- bai don vi KHONG THAY
    ho_so_sinh_that_...     FAILED   <- chi bai dau-cuoi bat duoc

Bai dau-cuoi kem hang doi chung: lo 300 seed phai co it nhat 20 ho so di dong,
khong thi khang dinh chua bao gio chay va bai xanh rong.

Xoa `parse_stringified_u8` — sau lan sua nay khong con ai goi.

Cong: fmt 0, clippy 0, test 0 (53/13/11 passed), msrv 1.88 -> 0.

Plane: VEIL-738

+++ fingerprint-core/src/types/fingerprint.rs
@@ -120,7 +120,11 @@ pub struct NavigatorFingerprint {
     /// `navigator.userAgent`
     pub user_agent: String,
     /// `navigator.hardwareConcurrency`
-    pub hardware_concurrency: u8,
+    /// So nhan logic. `u16` chu khong `u8`: mang Bayes khai bon gia tri
+    /// vuot 255 (384, 448, 512, 640), chiem 26,36% khoi luong xac suat —
+    /// `u8` lam chung roi vao `.unwrap_or(4)`, tuc mot con so TRONG HOP LY.
+    /// Xem VEIL-738.
+    pub hardware_concurrency: u16,
     /// `navigator.deviceMemory` — `None` when not exposed (Firefox, privacy budget).
     pub device_memory: Option<f32>,
     /// `navigator.platform`
